refactor(mcp): extract shared ingress and disable WSL computer use
- Add StreamableHttpMcpIngress for shared Streamable-HTTP MCP handling
- Delegate BrowserMcpIngress and ComputerUseMcpIngress to the shared helper
- Disable Computer Use for WSL projects in scope, composer, and thread UI
- Auto-clear computerUse config when scope becomes unavailable
- Add getComputerUseScope test for WSL project location
